fix(server): return a readable error message with 409 status code for integrity errors
Overridden the create function of the MedicalRecordViewSet to return a usable error message for the frontend to use on attempting to create an existing medical record of an appointment. Did the same thing for the PrescriptionViewSet to return the same error when attempting to create another prescription to the same medical record.